binary4.12

+++ pr_int.c
@@ -10,7 +10,7 @@
 int pr_int(va_list args, t_flag *fl)
 {
 	int i, j, num;
-	int x, y;
+	long int x, y;
 	char s[18] = {0};
 
 	x = va_arg(args, long int);
